Fix before/after type annotations to match their float query args
DbUndoSQLWeb.get_transactions() annotated before/after as int | None,
but TransactionsHistoryQueryArgs defines them as fields.Float and the
resource passes those float values straight through. No runtime change
(Python doesn't enforce annotations), just makes the signature match
actual usage.

+++ gramps_webapi/undodb.py
@@ -537,8 +537,8 @@ class DbUndoSQLWeb(DbUndoSQL):
         old_data: bool = True,
         new_data: bool = True,
         ascending: bool = True,
-        before: int | None = None,
-        after: int | None = None,
+        before: float | None = None,
+        after: float | None = None,
         before_id: int | None = None,
         after_id: int | None = None,
     ) -> tuple[list[dict[str, Any]], int]:
